Update UserGroup.php
Adds getUserLoginIDs and removerUser methods to UserGroup. The former returns an array of UserLoginIDs (not a user object) for a UserGroup; the other removes a single user with the provided loginID. This was needed to be able to circumvent the "remove all users, reinsert all users" processing previously handled in the Usergroup form.