v2025.04.03 Update (#813)
* Update admin.js

Fixes issue with the Subjects Admin list not showing up in the Resources Module.

* Updates Help Documentation Link

Updates link to use the Github Wiki and not Read the Docs.

* Update getAttachmentForm.php

Updates Attachment Form Submit button to be an input type button instead of submit to resolve form processing errors.

* Restores PR 703

PR 703 was accidentally overwritten in the 2025.04 Update; this commit restores the fix to Issue 687.

* Updates installer for 2025.04.03