v2025.10 (#831)
* Update Module-Specific Installers

Updates Module-Specific Installers to use default values rather than requiring additional CASE options for every new version. Now you will only need to update the module's installer if it's doing something (such as modifying the module's config file or running an SQL query). Refactors the module installer files further by moving configuration file parsing before the switch case. Removes vestigial commands in the module installer files (most often "have_read_write_access_to_config" being a dependency for an upgrade that doesn't actually write to the config file. Also fixes a situation in the 3.1.0 upgrade of the Licensing module where the configuration file was being rewritten for no reason.
 
* Removes installer module commentary

Commentary that applies to all module-specific installers was in only two of the files. Removed from those files and added to the Technical Documentation wiki on Github.

* Create versionManager Class

Creates a version manager class that centralizes a lot of the logic that was previously in index/installer.php. Adds method for getLastValidVersion and revises the installerVersion to be something decided by the last version number in the validVersionOrder list rather than something that has to be manually updated.

* Update installer.php

Adds upgrade string to the installer class.

* Create run_loop.php

Splits the run_loop function out from the install/index.php script. Updates error reporting in the script. Adds the version number to the upgrade message.

* Update index.php

Greatly reorganizes the install/index file in order to help make it a bit more readable. Still relies on session variables and is more of a functional program style, but at least uses more descriptive variable names and is structured to help the user follow what is happening. Hopefully this will make the advancement into more of an OOP installer a bit easier.

* Update db_tools.php

Adds mysqli_more_results to ensure additional results exist before checking the next result.

* Update set_installed_variable.php

Updates file to use versionManager to determine the installation version rather than a global variable in install/index.php

* Delete auth\install\index.php

Removes an empty index file to bring the module in alignment with the other modules.

* Update footers to dynamically update version numbers

Dynamically updates the footers to use the versionManager (for the main page) or config file (for module footers) to use the most-recently installed version of CORAL.

* Mild bugfix

Mild bugfix - sets the $_SESSION['ref_script'] *after* the $currentPage variable is set in the template/header.php file.